refactor: shared_preferences.dart のフォルダ移動
